Moved CRM Auth logic into module
